requirements.txt
apache-airflow-providers-amazon==5.1.0